Report gradual regression over time
The original design for this has a fundamental flaw, which is that
if performance drops by 0.9% every day, it will never report any
problems, even if at some point performance has dropped to a halt.

To fix this, we now search back until A. the first data point ever
recorded, or B. the last data point which was reported as regressing.

This means that the case of 0.9% will report on the second day, because
it will then see a decrease in performance of 1.8%.